:recycle: Move payment list to its own component
